Reduced unnecessary game reset execution

+++ sfall/LoadGameHook.cpp
@@ -86,25 +86,32 @@ void ClearLoopFlag(LoopFlag flag) {
 	inLoop &= ~flag;
 }
 
-static void _stdcall ResetState(DWORD onLoad) {
-	if (GraphicsMode > 3) Graphics_OnGameLoad();
-	ForceGraphicsRefresh(0); // disable refresh
-	LoadOrder_OnGameLoad();
-	WipeSounds();
-	Interface_OnGameLoad();
-	RestoreObjUnjamAllLocks();
-	Worldmap_OnGameLoad();
-	Stats_OnGameLoad();
-	PerksReset();
-	Combat_OnGameLoad();
-	Skills_OnGameLoad();
-	if (!onLoad) FileSystemReset();
-	InventoryReset();
-	PartyControl_OnGameLoad();
-	ResetExplosionRadius();
-	BarBoxes_OnGameLoad();
-	ScriptExtender_OnGameLoad();
+static void _stdcall ResetState(DWORD onLoad) { // OnGameReset
+	if (gameLoaded) { // prevent resetting when a new game has not been started (loading saved games from main menu)
+		if (GraphicsMode > 3) Graphics_OnGameLoad();
+		ForceGraphicsRefresh(0); // disable refresh
+		LoadOrder_OnGameLoad();
+		Interface_OnGameLoad();
+		RestoreObjUnjamAllLocks();
+		Worldmap_OnGameLoad();
+		Stats_OnGameLoad();
+		PerksReset();
+		Combat_OnGameLoad();
+		Skills_OnGameLoad();
+		FileSystemReset();
+		WipeSounds();
+		InventoryReset();
+		PartyControl_OnGameLoad();
+		ResetExplosionRadius();
+		BarBoxes_OnGameLoad();
+		ScriptExtender_OnGameLoad();
+		if (isDebug) {
+			char* str = (onLoad) ? "on Load" : "on Exit";
+			DebugPrintf("\n[SFALL: State reset %s]", str);
+		}
+	}
 	inLoop = 0;
+	gameLoaded = false;
 }
 
 void GetSavePath(char* buf, char* ftype) {
